number_recognition_system/requirements.txt
tensorflow==2.1.0
flask==1.1.2
jinja2==2.11.2
numpy==1.19.1
pillow==7.2.0